Seperate PickPlaceTask into new file
- Update profiler from benchmark overhaul
